pyproject.toml
[project]
name = "cs2-body-head-detection"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dvc>=3.65.0",
    "fire>=0.7.1",
    "hydra-core>=1.3.2",
    "kaggle>=1.8.3",
    "kagglehub>=0.3.13",
    "mlflow>=3.8.0",
    "omegaconf>=2.3.0",
    "onnx>=1.20.0",
    "pytorch-lightning>=2.6.0",
    "tensorboard>=2.20.0",
    "tensorboardx>=2.6.4",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "ultralytics>=8.3.241",
]

[dependency-groups]
dev = [
    "black>=25.12.0",
    "flake8>=7.3.0",
    "isort>=7.0.0",
    "pre-commit>=4.5.1",
]

[tool.isort]
profile = "black"
line_length = 88

[tool.black]
line-length = 88
